fix: move server export to its own file
